MyLibrary/requirement.txt
apturl==0.5.2
asttokens==2.4.1
awscli==1.22.34
awsebcli==3.20.9
Babel==2.8.0
bcrypt==4.2.0
blessed==1.20.0
blinker==1.4
botocore==1.31.43
Brlapi==0.8.3
cement==2.8.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
colorama==0.4.3
comm==0.2.2
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.5
decorator==5.1.1
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
docutils==0.17.1
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.6.0
Flask==2.0.1
Flask-JWT-Extended==4.6.0
fonttools==4.53.1
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.1.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==2.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
more-itertools==8.10.0
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==2.1.0
oauthlib==3.2.0
olefile==0.46
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pathspec==0.10.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==3.12.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==8.3.2
python-apt==2.4.0+ubuntu1
python-dateutil==2.9.0.post0
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.1.1
reportlab==3.6.8
requests==2.31.0
roman==3.3
rsa==4.8
s3transfer==0.5.0
SecretStorage==3.3.1
semantic-version==2.8.5
simplejson==3.17.6
six==1.16.0
stack-data==0.6.3
systemd-python==234
termcolor==1.1.0
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.1.9
Werkzeug==2.0.2
xdg==5
xkit==0.0.0
zipp==1.0.0
